pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sentinel"
version = "1.0.0"
description = "Defense Stock Claim Analyzer — classifies social media claims about defense stocks as exaggerated, accurate, or understated."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "twscrape",
    "spacy",
    "en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.8.0/en_core_web_lg-3.8.0-py3-none-any.whl",
    "openai",
    "google-genai",
    "python-dotenv",
    "pyyaml",
    "yfinance",
    "ddgs",
    "python-dateutil>=2.9.0.post0",
    "flask",
    "flask-cors",
    "psycopg[binary]",
    "scikit-learn",
    "transformers",
    "torch",
    "sentence-transformers",
    "matplotlib",
    "click",
    "vaderSentiment",
    "joblib",
    "gunicorn",
]

[project.scripts]
sentinel = "src.cli:cli"

[project.optional-dependencies]
local-embeddings = ["sentence-transformers"]

[dependency-groups]
dev = [
    "pylint",
    "pytest",
    "pytest-asyncio",
    "pytest-mock",
    "pytest-cov",
    "pytest-timeout",
    "freezegun",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
markers = [
    "slow: marks tests as slow",
    "integration: marks tests as integration tests",
]
timeout = 60

[tool.coverage.run]
source = ["src"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]
